4.2 done

+++ src/models/gpu_base.rs
@@ -1,9 +1,9 @@
 use serde::{Deserialize, Serialize};
 use sqlx::FromRow;
 
-#[derive(Debug, Serialize, Deserialize, FromRow)]
+#[derive(Debug, Clone, Serialize, Deserialize, FromRow)]
 pub struct GpuBase {
-    pub id: Option<i32>,
+    pub id: Option<i64>,
     pub name: String,
     pub brand: Option<String>,
 }
