+++ src/models/model_map.rs
@@ -1,9 +1,9 @@
 use serde::{Deserialize, Serialize};
 use sqlx::FromRow;
 
-#[derive(Debug, Serialize, Deserialize, FromRow)]
+#[derive(Debug, Clone, Serialize, Deserialize, FromRow)]
 pub struct ModelMap {
-    pub id: Option<i32>,
+    pub id: Option<i64>,
     pub model_name: Option<String>,
     pub base_model: Option<String>,
 }
